moved dcircle to shared talents and added number of tp's
