+++ src/analysis/retail/warlock/shared/index.ts
@@ -5,4 +5,4 @@ export { SoulShardIcon };
 export { default as GrimoireOfSacrifice } from './GrimoireOfSacrifice';
 export { default as GrimoireOfSacrificeNormalizer } from './GrimoireOfSacrificeNormalizer';
 export { default as WarlockMissingDotApplyDebuffPrePull } from './WarlockMissingDotApplyDebuffPrePull';
-export { default as DemonicCirclesCreated } from './DemonicCirclesCreated';
+export { default as DemonicCircle } from './talents/DemonicCircle';
